Added Values function to project.
