+++ cmd/internal/run.go
@@ -1,9 +1,13 @@
 package internal
 
 import (
+	"regexp"
+	"strings"
 	"io"
 
 	"github.com/bilus/oya/pkg/project"
+	"github.com/bilus/oya/pkg/template"
+
 )
 
 func Run(workDir, taskName string, positionalArgs []string, flags map[string]string, stdout, stderr io.Writer) error {
@@ -11,6 +15,35 @@ func Run(workDir, taskName string, positionalArgs []string, flags map[string]str
 	if err != nil {
 		return err
 	}
+	values, err := p.Values()
+	if err != nil {
+		return err
+	}
+	return p.Run(workDir, taskName, toScope(positionalArgs, flags).Merge(values), stdout, stderr)
+}
 
-	return p.Run(workDir, taskName, positionalArgs, flags, stdout, stderr)
+func toScope(positionalArgs []string, flags map[string]string) template.Scope {
+	return template.Scope{
+		"Args":  positionalArgs,
+		"Flags": camelizeFlags(flags),
+	}
+}
+
+func camelizeFlags(flags map[string]string) map[string]string {
+	result := make(map[string]string)
+	for k, v := range flags {
+		result[camelize(k)] = v
+	}
+	return result
+}
+
+var sepRx = regexp.MustCompile("(-|_).")
+
+// camelize turns - or _ separated identifiers into camel case.
+// Example: "aa-bb" becomes "aaBb".
+func camelize(s string) string {
+	return sepRx.ReplaceAllStringFunc(s, func(match string) string {
+		return strings.ToUpper(match[1:])
+	})
+
 }
